+++ tests/runner_test.py
@@ -14,14 +14,6 @@ class TestRunner(object):
 		assert 'disk' in system_test
 		assert 'loadavg' in system_test
 
-	def test_system_empty(self):
-		empty = runner.empty()
-		if sys.platform != 'darwin':
-			assert 'network' in empty
-		assert 'memory' in empty
-		assert 'cpu' in empty
-		assert 'disk' in empty
-		assert 'loadavg' in empty
 
 	def test_process_run(self):
 		processes = runner.processes()