+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+1.0
+======
+
+* Improved memory usage stats
+* Gets Linux distribution information
+
+
 0.3.4
 -----
 Fixed locale